Add support for registering routes by regex
Changes to NativeFunctionAdapter and FunctionManager to support
registering function with regex routes. Changes do not alter existing
functionality, just lay ground work for future lookup.

NativeFunctionAdapter objects now have optional constructor arguments of
route and action (actions are HTTP actions). If not specified, route is
set to ^function.name$ and action is set to all actions. So default
action should match current implementation.

FixedFunctionManager is altered to have an alternate storage of routes
in an ordered list. This means that the order of registration matters,
and the first match will be run.