SOAP Fixes
Change add_function to always register simple function name routes for
GET and POST so that SOAP finds them. They are lower in the list the
user defined routes for REST, so they won't be called.
Change get_all to use the new data structures to return a unique list of
all functions.